% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1blas.R
\name{amaxgpu}
\alias{amaxgpu}
\title{amaxgpu}
\usage{
amaxgpu(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
the resulting index
}
\description{
This function finds the (smallest) index of the element with 
the maximum magnitude of given vector/matrix
by using CUDA cublas function cublasIdamax
}
\examples{
a <- 1:4
a_gpu <- creategpu(a)
amaxgpu(a_gpu)
}
\seealso{
\code{\link{amingpu}}
}

