% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1blas.R
\name{asumgpu}
\alias{asumgpu}
\title{asumgpu}
\usage{
asumgpu(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
the vector/matrix's elements absolute values summation
}
\description{
This function computes the summation 
of the absolute values of elements of given vector/matrix
by using CUDA cublas function cublasDasum
}
\examples{
a <- 1:4
a_gpu <- creategpu(a)
asumgpu(a_gpu) 
}
\seealso{
\code{\link{maxgpu}}
}

