% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1blas.R
\name{axpygpu}
\alias{axpygpu}
\title{axpygpu}
\usage{
axpygpu(x, y, alpha = 1)
}
\arguments{
\item{x}{list consisting of R external GPU pointer and dimension}

\item{y}{list consisting of R external GPU pointer and dimension}

\item{alpha}{scale factor alpha; default 1}
}
\value{
updated y vector/matrix
}
\description{
This function multiplies the vector x by the scalar a and adds it 
to the vector y overwriting the latest vector with the result 
by using CUDA cublas function cublasDaxpy
}
\examples{
a <- 1:4
a_gpu <- creategpu(a)
b_gpu <- creategpu(a)
axpygpu(a_gpu, b_gpu, 1) 
}
\seealso{
\code{\link{scalegpu}}
}

