% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_wise_math.R
\name{betaGPU}
\alias{betaGPU}
\title{betaGPU}
\usage{
betaGPU(input1, input2)
}
\arguments{
\item{input}{(non-negative vector) list consisting of R external GPU pointer and dimension}
}
\value{
beta result, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function apply the beta function to a given vector or matrix
by using CUDA function
}
\examples{
a <- 1:4
a_gpu <- createGPU(a)
betaGPU(a_gpu)->b_gpu
gatherGPU(b_gpu)
}
\seealso{
\code{\link{expGPU}}
}

