% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1blas.R
\name{copygpu}
\alias{copygpu}
\title{copygpu}
\usage{
copygpu(x, y)
}
\arguments{
\item{x}{list consisting of R external GPU pointer and dimension}

\item{y}{list consisting of R external GPU pointer and dimension}
}
\value{
copied vector/matrix
}
\description{
This function copies the vector x into the vector y  
by using CUDA cublas function cublasDcopy
}
\examples{
a <- 1:4
b <- 2:5
a_gpu <- creategpu(a)
b_gpu <- creategpu(b)
copygpu(a_gpu, b_gpu) 
}
\seealso{
\code{\link{axpygpu}}
}

