% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_object_interface.R
\name{creategpu}
\alias{creategpu}
\title{creategpu}
\usage{
creategpu(input, nrow = NULL, ncol = NULL)
}
\arguments{
\item{input}{R vector to be copied}

\item{nrow}{the desired number of rows}

\item{ncol}{the desired number of columns}
}
\value{
a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
Create a GPU vector/matrix by copying from the input R vector
}
\details{
This function creates a vector/matrix in GPU by calling the CUDA 
cudamalloc function, and then copys from input R vector.
The output of this function is a list consisting of the GPU pointer and 
its dimension.

If either one of nrow or ncol is not given, an one column matrix/vector is 
returned. This function returns row-major matrix.
}
\note{
output is a R external GPU pointer and can only be used in Rcublas functions
}
\examples{
a <- rnorm(6)
a_gpu <- creategpu(a, 2, 3)
gathergpu(a_gpu) 
}
\author{
Yuan Li
}
\seealso{
\code{\link{gathergpu}}
}
\keyword{GPU}
\keyword{create}
\keyword{vector}

