% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGPUmat.R
\name{createGPUmat}
\alias{createGPUmat}
\title{createGPUmat}
\usage{
createGPUmat(input, nrow, ncol)
}
\arguments{
\item{input}{R vector (matrix stored in column-major format) to be copied (numerical)}

\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\value{
a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
Create a GPU matrix by copying from the input R vector
}
\details{
This function creates a matrix in GPU by calling the CUDA cudamalloc function, and then copys
the values of matrix from input R vector (matrix stored in column-major format). The output is a list consisting of the GPU pointer and matrix dimension
}
\note{
output is a R external GPU pointer and can only be used in Rcublas functions
}
\examples{
a <- 1:4
am_gpu <- createGPUmat(a, 2, 2)
gatherGPU(am_gpu) 
}
\author{
Yuan Li
}
\seealso{
\code{\link{gatherGPU}} \code{\link{createGPU}}
}
\keyword{GPU}
\keyword{create}
\keyword{matrix}

