\name{cudaEventCreate}
\alias{cudaEventCreate}
\title{Creates an event object}
\description{ Creates an event object using \code{cudaEventDefault}.}
\usage{cudaEventCreate()}

\value{event}
\seealso{\code{\link{cudaEventCreate}}
\code{\link{cudaEventCreateWithFlags}}
\code{\link{cudaEventRecord}}
\code{\link{cudaEventQuery}}
\code{\link{cudaEventSynchronize}}
\code{\link{cudaEventDestroy}}
\code{\link{cudaEventElapsedTime}}
\code{\link{cudaStreamWaitEvent}}}
\references{\url{http://docs.nvidia.com/cuda/cuda-driver-api/index.htm}}
\keyword{programming}
\concept{GPU}
