\name{cudaGetDevice}
\alias{cudaGetDevice}
\title{Returns which device is currently being used}
\description{ Returns  the current device for the calling host thread.}
\usage{
cudaGetDevice()
}
\value{device}
\seealso{\code{cudaGetDeviceCount}
\code{cudaSetDevice}
\code{cudaGetDeviceProperties}
\code{\link{cudaChooseDevice}}}
\references{\url{http://docs.nvidia.com/cuda/cuda-driver-api/index.htm}}
\keyword{programming}
\concept{GPU}
