\name{cudaStreamCreate}
\alias{cudaStreamCreate}
\title{Create an asynchronous stream}
\description{ Creates a new asynchronous stream.}
\usage{cudaStreamCreate()}

\value{pStream}
\seealso{\code{\link{cudaStreamCreate}}
\code{\link{cudaStreamCreateWithFlags}}
\code{\link{cudaStreamQuery}}
\code{\link{cudaStreamSynchronize}}
\code{\link{cudaStreamWaitEvent}}
\code{\link{cudaStreamAddCallback}}
\code{\link{cudaStreamDestroy}}}
\references{\url{http://docs.nvidia.com/cuda/cuda-driver-api/index.html}}
\keyword{programming}
\concept{GPU}
