\name{cuInit}
\alias{cuInit}
\title{Initialize CUDA}
\description{
  This function ensures that the CUDA SDK is initialized.
}
\usage{
cuInit(flags = 0L)
}
\arguments{
  \item{flags}{an integer vector giving the configuration parameters for
  starting the CUDA SDK}
}
\value{
}
\references{}
\author{Duncan Temple Lang}
\seealso{}
\examples{
}
\keyword{programming}
