% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_wise_math.R
\name{dgammagpu}
\alias{dgammagpu}
\title{dgammagpu}
\usage{
dgammagpu(input, k = 1, theta = 1)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}

\item{k}{shape parameter of Gamma distribution; default value 1}

\item{theta}{scale parameter of Gamma distribution; default value 1}
}
\value{
gamma pdf result of vector/matrix, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the gammma pdf function of given vector/matrix
by using self-defined CUDA function
}
\examples{
a <- 1:4
a_gpu <- creategpu(a)
dgammagpu(a_gpu) -> b_gpu
gathergpu(b_gpu)
}
\seealso{
\code{\link{dbetagpu}}
}

