% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lapack.R
\name{dgmmgpu}
\alias{dgmmgpu}
\title{dgmmgpu}
\usage{
dgmmgpu(sidemode = 1, A, x, C)
}
\arguments{
\item{sidemode}{indicates whether the given matrix is on the left or right side
in the matrix equation solved by a particular function. If sidemode == 1, 
the matrix is on the left side in the equation If sidemode == 2, 
the matrix is on the right side in the equation.}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{x}{input vector; list of R external GPU pointer and dimension}

\item{C}{input/output matrix; list of R external GPU pointer and dimension}
}
\value{
updated matrix C, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix C's number of rows}
\item{n: }{matrix C's number of columns}
}
}
\description{
This function performs the matrix-matrix multiplication 
C = A diag(x) or C = diag(x) A 
by using CUDA cublas function cublasDdgmm
}
\seealso{
\code{\link{symmgpu}}
}

