% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{dnormgpu}
\alias{dnormgpu}
\title{dnormgpu}
\usage{
dnormgpu(input, mean = 0, sd = 1)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}

\item{mean}{vector/matrix of mean}

\item{sd}{vector/matrix of standard deviation}
}
\value{
normal distribution density vector/matrix, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the normal distribution density of given vector/matrix
}
\details{
If mean or sd are not specified they assume the default values
of 0 and 1, respectively.
}
\examples{
a <- 1:4
a_gpu <- creategpu(a)
dnormgpu(a_gpu) -> b_gpu
gathergpu(b_gpu)
}
\seealso{
\code{\link{pnormgpu}}
}

