% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expGPU.R
\name{expGPU}
\alias{expGPU}
\title{expGPU}
\usage{
expGPU(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
exponential vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the exponential of given vector or matrix
by using CUDA function
}
\examples{
a <- 1:4
a_gpu <- createGPU(a)
expGPU(a_gpu)->b_gpu
gatherGPU(b_gpu)
}
\seealso{
\code{\link{scaleGPU}}
}

