% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherGPU.R
\name{gatherGPU}
\alias{gatherGPU}
\title{gatherGPU}
\usage{
gatherGPU(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
R vector
}
\description{
Copy GPU vector to R vector
}
\details{
This function copys GPU vector/matrix to R vector
}
\note{
output is R vector and can be used by any R functions
}
\examples{
a <- 1:4
am_gpu <- createGPUmat(a, 2, 2)
gatherGPU(am_gpu) 
}
\author{
Yuan Li
}
\seealso{
\code{\link{gatherGPU}} \code{\link{createGPU}}
}
\keyword{GPU}

