% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{gbmvgpu}
\alias{gbmvgpu}
\title{gbmvgpu}
\usage{
gbmvgpu(A, x, y, kl, ku, alpha = 1, beta = 0, trans = 1)
}
\arguments{
\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{x}{input vector; list of R external GPU pointer and dimension}

\item{y}{input/output vector; list of R external GPU pointer and dimension}

\item{kl}{number of subdiagonals}

\item{ku}{number of superdiagonals}

\item{alpha}{scale factor a of banded matrix A; default 1}

\item{beta}{scale factor b of vector y; default 0}

\item{trans}{matrix A transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}
}
\value{
vector y, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{length of vector y}
\item{n: }{1}
}
}
\description{
This function computes banded matrix-vector multipication y = a A x + b y
by using CUDA cublas function cublasDgbmv
}
\examples{
A <- 1:12
x <- 1:3
y <- 1:4
A_gpu <- creategpu(A, 4, 3)
x_gpu <- creategpu(x)
y_gpu <- creategpu(y)
gemvgpu(A_gpu, x_gpu, y_gpu, 1, 1, 1)
gathergpu(y_gpu)
}
\seealso{
\code{\link{gergpu}}
}

