% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3blas.R
\name{gemmgpu}
\alias{gemmgpu}
\title{gemmgpu}
\usage{
gemmgpu(A, B, C, transa = 1, transb = 1, alpha = 1, beta = 0)
}
\arguments{
\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{B}{input matrix; list of R external GPU pointer and dimension}

\item{C}{input/output matrix; list of R external GPU pointer and dimension}
}
\value{
updated matrix C, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix X's number of rows}
\item{n: }{matrix Y's number of columns}
}
}
\description{
This function computes the matrix-matrix multiplication 
C = a op ( A ) op ( B ) + b C 
by using CUDA cublas function cublasDgemm
}
\examples{
a <- 1:6
b <- 2:7
a_gpu <- creategpu(a, 2, 3)
b_gpu <- creategpu(b, 3, 2)
mmgpu(a_gpu, b_gpu) -> c_gpu
gathergpu(c_gpu)
}
\seealso{
\code{\link{mmgpu}}
}

