% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{gemvgpu}
\alias{gemvgpu}
\title{gemvgpu}
\usage{
gemvgpu(trans = 1, alpha = 1, A, x, beta = 0, y)
}
\arguments{
\item{trans}{matrix A transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}

\item{alpha}{scale factor a of matrix A; default 1}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{x}{input vector; list of R external GPU pointer and dimension}

\item{beta}{scale factor b of vector y; default 0}

\item{y}{input/output vector; list of R external GPU pointer and dimension}
}
\value{
vector y, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{length of vector y}
\item{n: }{1}
}
}
\description{
This function computes matrix-vector multipication y = a A x + b y
by using CUDA cublas function cublasDgemv
}
\examples{
A <- 1:12
x <- 1:3
y <- 1:4
A_gpu <- creategpu(A, 4, 3)
x_gpu <- creategpu(x)
y_gpu <- creategpu(y)
gemvgpu(trans = 1, alpha = 1, A_gpu, x_gpu, beta = 1, y_gpu)
gathergpu(y_gpu)
}
\seealso{
\code{\link{gergpu}}
}

