% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{gergpu}
\alias{gergpu}
\title{gergpu}
\usage{
gergpu(A, x, y, alpha = 1)
}
\arguments{
\item{A}{input/output matrix; list of R external GPU pointer and dimension}

\item{x}{input vector; list of R external GPU pointer and dimension}

\item{y}{input vector; list of R external GPU pointer and dimension}

\item{alpha}{scale factor a of matrix A; default 1}
}
\value{
updated matrix A, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix A's number of rows}
\item{n: }{matrix A's number of columns}
}
}
\description{
This function perform the the rank-1 update A = a x y T + A,
by using CUDA cublas function cublasDger
}
\examples{
A <- 1:12
x <- 1:3
y <- 1:4
A_gpu <- creategpu(A, 3, 4)
x_gpu <- creategpu(x)
y_gpu <- creategpu(y)
gergpu(A_gpu, x_gpu, y_gpu, 1)
gathergpu(A_gpu)
}
\seealso{
\code{\link{gemvgpu}}
}

