% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_object_class.R
\name{GPUobject}
\alias{GPUobject}
\title{GPUobject}
\usage{
GPUobject(input, nrow, ncol)
}
\arguments{
\item{input}{R external pointer}

\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\value{
a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
classify the input as GPU vector/matrix and assign its dimension
}
\details{
This function classifies the input object as GPU vector/matrix
and assign its dimension
The output of this function is a list consisting of the GPU pointer
and its dimension
}
\note{
output is a R external GPU pointer and can only be used in Rcublas functions
}
\author{
Yuan Li
}
\seealso{
\code{\link{gathergpu}}
}
\keyword{GPU}
\keyword{create}
\keyword{vector}

