% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_query.R
\name{gpuquery}
\alias{gpuquery}
\title{gpuquery
 
This function returns the information of available GPU device in system}
\usage{
gpuquery()
}
\description{
gpuquery
 
This function returns the information of available GPU device in system
}
\examples{
gpuquery()
}
\seealso{
\code{\link{creategpu}}
}

