% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lapack.R
\name{inversegpu}
\alias{inversegpu}
\title{inversegpu}
\usage{
inversegpu(X)
}
\arguments{
\item{X}{input matrix; list of R external GPU pointer and dimension}
}
\value{
matrix inversion, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the inversion of given matrix (squared) 
by using CUDA cublas function cublasDgetrfBatched 
and cublasDgetriBatched (LU decomposition)
}
\examples{
a <- 1:9
a_gpu <- creategpu(a, 3, 3)
inversegpu(a_gpu) -> c_gpu
gathergpu(c_gpu)
}
\seealso{
\code{\link{mmgpu}} \code{\link{creategpu}}
}

