% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_wise_math.R
\name{logGPU}
\alias{logGPU}
\title{logGPU}
\usage{
logGPU(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
natural logarithms vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the natural logarithms of given vector or matrix
by using CUDA function
}
\examples{
a <- 1:4
a_gpu <- createGPU(a)
logGPU(a_gpu)->b_gpu
gatherGPU(b_gpu)
}
\seealso{
\code{\link{expGPU}}
}

