% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlognormGPU.R
\name{lognormRNGGPU}
\alias{lognormRNGGPU}
\title{log normal random number generator}
\usage{
lognormRNGGPU(n, mean = 0, sd = 1, seed = 1)
}
\arguments{
\item{n}{number of random numbers}

\item{mean}{mean of log-normal distribution; default value 0}

\item{sd}{standard deviation of log-normal distribution; default value 1}

\item{seed}{random number generator seed}
}
\value{
random numbers vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix X's number of rows}
\item{n: }{matrix X's number of columns}
}
}
\description{
This function generates log-normally distributed numbers 
by using CUDA curand function CURAND_RNG_PSEUDO_DEFAULT and curandGenerateLogNormalDouble
}
\examples{
a_gpu <- lognormRNGGPU(100, 0, 1, 15)
gatherGPU(a_gpu)
}
\seealso{
\code{\link{normRNGGPU}}
}

