% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1blas.R
\name{maxgpu}
\alias{maxgpu}
\title{maxgpu}
\usage{
maxgpu(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
the resulting index
}
\description{
This function finds the (smallest) index of the element with 
the maximum magnitude of given vector/matrix
by using CUDA cublas function cublasIdamin
}
\examples{
a <- 1:4
a_gpu <- creategpu(a)
maxgpu(a_gpu)
}
\seealso{
\code{\link{mingpu}}
}

