% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_define_CUDA_functions.R
\name{meangpu}
\alias{meangpu}
\title{meangpu}
\usage{
meangpu(x)
}
\arguments{
\item{x}{list consisting of R external GPU pointer and dimension}
}
\value{
vector/matrix mean
}
\description{
Compute the mean of given vector/matrix
}
\details{
This function computes the mean of given vector/matrix 
by using self-defined CUDA function
}
\examples{
a <- creategpu(1:4)
meangpu(a)
}
\author{
Yuan Li
}
\seealso{
\code{\link{sumgpu}}
}
\keyword{GPU}

