% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1blas.R
\name{mingpu}
\alias{mingpu}
\title{mingpu}
\usage{
mingpu(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
the resulting index
}
\description{
This function finds the (smallest) index of the element 
with the minimum magnitude of given vector
by using CUDA cublas function cublasIdamin
}
\examples{
a <- 1:4
a_gpu <- creategpu(a)
mingpu(a_gpu) 
}
\seealso{
\code{\link{maxgpu}}
}

