% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3blas.R
\name{mmGPU}
\alias{mmGPU}
\title{mmGPU}
\usage{
mmGPU(X, Y)
}
\arguments{
\item{X}{input matrix; list of R external GPU pointer and dimension}

\item{Y}{input matrix; list of R external GPU pointer and dimension}
}
\value{
matrix-matrix multiplication (X * Y), a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix X's number of rows}
\item{n: }{matrix Y's number of columns}
}
}
\description{
This function computes the matrix-matrix multiplication (X * Y) 
by using CUDA cublas function cublasDgemm
}
\examples{
a <- 1:6
b <- 2:7
a_gpu <- createGPUmat(a,2,3)
b_gpu <- createGPUmat(b,3,2)
mmGPU(a_gpu, b_gpu)->c_gpu
gatherGPU(c_gpu)
}
\seealso{
\code{\link{mmGPU}} \code{\link{createGPUmat}}
}

