% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{mvgpu}
\alias{mvgpu}
\title{mvgpu}
\usage{
mvgpu(X, y)
}
\arguments{
\item{X}{input matrix; list of R external GPU pointer and dimension}

\item{y}{input vector; list of R external GPU pointer and dimension}
}
\value{
matrix-vector multiplication (X * y), a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix X's number of rows}
\item{n: }{matrix X's number of columns; vector y's number of elements}
}
}
\description{
This function computes the matrix-vector multiplication (X * y) 
by using CUDA cublas function cublasDgemv
}
\examples{
a <- 1:4
b <- 2:3
a_gpu <- creategpu(a, 2, 2)
b_gpu <- creategpu(b)
mvgpu(a_gpu, b_gpu) -> c_gpu
gathergpu(c_gpu)
}
\seealso{
\code{\link{mmgpu}}
}

