% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1blas.R
\name{norm2gpu}
\alias{norm2gpu}
\title{norm2gpu}
\usage{
norm2gpu(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
vector Euclidean norm, a non-negative number
}
\description{
This function computes Euclidean norm of given 
vector/matrix by using CUDA cublas function cublasDnrm2
}
\examples{
a <- 1:4
a_gpu <- creategpu(a)
norm2gpu(a_gpu) 
}
\author{
Yuan Li
}
\seealso{
\code{\link{gathergpu}}
}
\keyword{GPU}

