% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{pnormgpu}
\alias{pnormgpu}
\title{pnormgpu}
\usage{
pnormgpu(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
standard normal CDF, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the standard normal distribution cumulative 
density (CDF) of given vector/matrix
}
\examples{
a <- 1:4
a_gpu <- creategpu(a)
pnormgpu(a_gpu) -> b_gpu
gathergpu(b_gpu)
}
\seealso{
\code{\link{dnormgpu}}
}

