% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number_genertor.R
\name{poissonRNGGPU}
\alias{poissonRNGGPU}
\title{Poisson random number generator}
\usage{
poissonRNGGPU(n, lambda = 1, seed = 1)
}
\arguments{
\item{n}{number of random numbers}

\item{lambda}{mean of Poisson distribution; default value 1}

\item{seed}{random number generator seed}
}
\value{
random numbers vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix X's number of rows}
\item{n: }{matrix X's number of columns}
}
}
\description{
This function generates Poisson distributed numbers 
by using CUDA curand function CURAND_RNG_PSEUDO_DEFAULT and curandGeneratePoisson
}
\examples{
a_gpu <- poissonRNGGPU(100, 1)
gatherGPU(a_gpu)
}
\seealso{
\code{\link{normRNGGPU}}
}

