% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerGPU.R
\name{powerGPU}
\alias{powerGPU}
\title{powerGPU}
\usage{
powerGPU(input, alpha)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}

\item{alpha}{power factor}
}
\value{
powered vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the power of given vector or matrix 
by using CUDA function
}
\examples{
a <- 1:4
b <- 2
a_gpu <- createGPU(a)
powerGPU(a_gpu, b)->b_gpu
gatherGPU(b_gpu)
}
\seealso{
\code{\link{sqrtGPU}}
}

