\name{Rcublas}
\docType{package}
\title{
GPU enabled BLAS functions and random number generators

}
\description{
This package provides GPU-accelerated algebra and random numbergenerator functions by wrapping CUDA library. It also includes some self-defined GPU algebra functions which are unavailable from CUDA library.}
\details{
\tabular{ll}{
Package: \tab Rcublas\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-01-07\cr
License: \tab CPL\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Yuan Li, Hua Zhou

Maintainer: Yuan Li <yli16@ncsu.edu>}
 

\keyword{GPU, CUDA, Passing by Reference }
 