% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number_genertor.R
\name{rgammagpu}
\alias{rgammagpu}
\title{rgammagpu}
\usage{
rgammagpu(n, alpha = 2, beta = 1, seed = 1)
}
\arguments{
\item{n}{number of random numbers}

\item{alpha}{par of Gamma distribution; default value 2}

\item{beta}{par of Gamma distribution; default value 1}

\item{seed}{random number generator seed; default value 1}
}
\value{
generated random numbers vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function generates Gamma distributed random numbers 
by using self-defined CUDA function based on George Marsaglia 
and Wai Wan Tsang's method
}
\examples{
a_gpu <- rgammagpu(100, 2, 1) 
}
\seealso{
\code{\link{runifgpu}}
}

