% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number_genertor.R
\name{rpoisgpu}
\alias{rpoisgpu}
\title{rpoisgpu}
\usage{
rpoisgpu(n, lambda = 1, seed = 1)
}
\arguments{
\item{n}{number of random numbers}

\item{lambda}{mean of Poisson distribution; default value 1}

\item{seed}{random number generator seed; default value 1}
}
\value{
generated random numbers vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function generates Poisson distributed random numbers 
by using CUDA curand function CURAND_RNG_PSEUDO_DEFAULT 
and curandGeneratePoisson
}
\examples{
a_gpu <- rpoisgpu(100, 1) 
}
\seealso{
\code{\link{runifgpu}}
}

