% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number_genertor.R
\name{runifgpu}
\alias{runifgpu}
\title{runifgpu}
\usage{
runifgpu(n, seed = 1)
}
\arguments{
\item{n}{number of random numbers}

\item{seed}{random number generator seed; default value 1}
}
\value{
generated random numbers vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function generates uniformly distributed random numbers between 0 and 1 
by using CUDA curand function CURAND_RNG_PSEUDO_DEFAULT 
and curandGenerateUniformDouble
}
\examples{
a_gpu <- runifgpu(100, 15)
gathergpu(a_gpu)
}
\seealso{
\code{\link{creategpu}}
}

