% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{sbmvgpu}
\alias{sbmvgpu}
\title{sbmvgpu}
\usage{
sbmvgpu(fillmode = 1, k, alpha = 1, A, x, beta = 0, y)
}
\arguments{
\item{fillmode}{indicates if matrix A lower or upper part is stored, 
the other symmetric part is not referenced and is inferred from the 
stored elements. if fillmode == 1 then the symmetric 
banded matrix A is stored column by column, with the main diagonal of 
the matrix stored in row 1, the first subdiagonal in row 2 
(starting at first position), 
the second subdiagonal in row 3 (starting at first position), etc. 
if fillmode == 2 then the symmetric banded matrix A is 
stored column by column, with the main diagonal of the matrix stored 
in row k+1, the first superdiagonal in row k (starting at second position),
the second superdiagonal in row k-1 (starting at third position), etc.}

\item{k}{number of subdiagonals}

\item{alpha}{scale factor a of symmetric banded matrix A; default 1}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{x}{input vector; list of R external GPU pointer and dimension}

\item{beta}{scale factor b of vector y; default 0}

\item{y}{input/output vector; list of R external GPU pointer and dimension}
}
\value{
vector y, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{length of vector y}
\item{n: }{1}
}
}
\description{
This function computes symmetric banded matrix-vector multipication y = a A x + b y
by using CUDA cublas function cublasDsbmv
}
\seealso{
\code{\link{gemvgpu}}
}

