% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleGPU.R
\name{scaleGPU}
\alias{scaleGPU}
\title{scaleGPU}
\usage{
scaleGPU(input, alpha)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}

\item{alpha}{scale factor}
}
\value{
scaled vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function scales the given vector by a scalar
by using CUDA cublas function cublasDcopy
}
\examples{
a <- 1:4
b <- 2
a_gpu <- createGPU(a)
scaleGPU(a_gpu, b)->b_gpu
gatherGPU(b_gpu)
}
\seealso{
\code{\link{expGPU}}
}

