% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_define_CUDA_functions.R
\name{subsetGPU}
\alias{subsetGPU}
\title{subsetGPU}
\usage{
subsetGPU(input, index)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}

\item{index}{index of the vector subset}
}
\value{
subset of vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function select and copy subset of a given GPU vector
by using CUDA function
}
\examples{
a <- 1:4
a_gpu <- createGPU(a)
subsetGPU(a_gpu,c(1,2))->b_gpu
gatherGPU(b_gpu)
}
\seealso{
\code{\link{createGPU}}
}

