% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_define_CUDA_functions.R
\name{sumGPU}
\alias{sumGPU}
\title{sumGPU}
\usage{
sumGPU(x)
}
\arguments{
\item{x}{list consisting of R external GPU pointer and dimension}
}
\value{
vector sum
}
\description{
Compute the sum of given vector
}
\details{
This function computes sum of given vector by using CUDA vector reduction
}
\examples{
a <- createGPU(1:4)
sumGPU(a)
}
\author{
Yuan Li
}
\seealso{
\code{\link{gatherGPU}} \code{\link{createGPU}}
}
\keyword{GPU}

