% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3blas.R
\name{symmgpu}
\alias{symmgpu}
\title{symmgpu}
\usage{
symmgpu(sidemode = 1, fillmode = 1, alpha = 1, A, B, beta = 0, C)
}
\arguments{
\item{sidemode}{indicates whether the given matrix is on the left or right side
in the matrix equation solved by a particular function. If sidemode == 1, 
the matrix is on the left side in the equation If sidemode == 2, 
the matrix is on the right side in the equation.}

\item{fillmode}{indicates if matrix A lower or upper part is stored, 
the other part is not referenced and is inferred from the 
stored elements. if fillmode == 1 then the triagular 
banded matrix A is stored column by column, with the main diagonal of 
the matrix stored in row 1, the first subdiagonal in row 2 
(starting at first position), 
the second subdiagonal in row 3 (starting at first position), etc. 
if fillmode == 2 then the triangular banded matrix A is 
stored column by column, with the main diagonal of the matrix stored 
in row k+1, the first superdiagonal in row k (starting at second position),
the second superdiagonal in row k-1 (starting at third position), etc.}

\item{alpha}{scale factor a of matrix AB; default 1}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{B}{input matrix; list of R external GPU pointer and dimension}

\item{beta}{scale factor b of matrix C; default 0}

\item{C}{input/output matrix; list of R external GPU pointer and dimension}
}
\value{
updated matrix C, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix C's number of rows}
\item{n: }{matrix C's number of columns}
}
}
\description{
This function computes the symmetric matrix-matrix multiplication 
C = a A B + b C 
by using CUDA cublas function cublasDsymm
}
\seealso{
\code{\link{gemmgpu}}
}

