% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{symvgpu}
\alias{symvgpu}
\title{symvgpu}
\usage{
symvgpu(fillmode = 1, alpha = 1, A, x, beta = 0, y)
}
\arguments{
\item{fillmode}{indicates if matrix A lower or upper part is stored, 
the other symmetric part is not referenced and is inferred from the 
stored elements. if fillmode == 1 then the symmetric 
banded matrix A is stored in lower mode
if fillmode == 2 then the symmetric banded matrix A is 
stored in upper mode}

\item{alpha}{scale factor a of symmetric banded matrix A; default 1}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{x}{input vector; list of R external GPU pointer and dimension}

\item{beta}{scale factor b of vector y; default 0}

\item{y}{input/output vector; list of R external GPU pointer and dimension}
}
\value{
vector y, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{length of vector y}
\item{n: }{1}
}
}
\description{
This function computes symmetric matrix-vector multipication y = a A x + b y
by using CUDA cublas function cublasDsymv
}
\seealso{
\code{\link{sbmvgpu}}
}

