% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3blas.R
\name{syr2kgpu}
\alias{syr2kgpu}
\title{syr2kgpu}
\usage{
syr2kgpu(fillmode = 1, trans = 1, alpha = 1, A, B, beta = 0, C)
}
\arguments{
\item{fillmode}{indicates if matrix A lower or upper part is stored, 
the other part is not referenced and is inferred from the 
stored elements. if fillmode == 1 then the triagular 
banded matrix A is stored column by column, with the main diagonal of 
the matrix stored in row 1, the first subdiagonal in row 2 
(starting at first position), 
the second subdiagonal in row 3 (starting at first position), etc. 
if fillmode == 2 then the triangular banded matrix A is 
stored column by column, with the main diagonal of the matrix stored 
in row k+1, the first superdiagonal in row k (starting at second position),
the second superdiagonal in row k-1 (starting at third position), etc.}

\item{trans}{matrix A and B transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}

\item{alpha}{scale factor a ; default 1}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{B}{input matrix; list of R external GPU pointer and dimension}

\item{beta}{scale factor b; default 0}

\item{C}{input/output matrix; list of R external GPU pointer and dimension}
}
\value{
updated matrix C, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix C's number of rows}
\item{n: }{matrix C's number of columns}
}
}
\description{
This function performs the symmetric rank- 2k update 
C = a(op ( A )op ( B ) T + op ( B )op ( A )T)  + b C
by using CUDA cublas function cublasDsyr2k
}
\seealso{
\code{\link{syrkgpu}}
}

