% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{syrgpu}
\alias{syrgpu}
\title{syrgpu}
\usage{
syrgpu(fillmode = 1, alpha = 1, x, A)
}
\arguments{
\item{fillmode}{indicates if matrix A lower or upper part is stored, 
the other symmetric part is not referenced and is inferred from the 
stored elements. if fillmode == 1 then the symmetric 
banded matrix A is stored in lower mode
if fillmode == 2 then the symmetric banded matrix A is 
stored in upper mode}

\item{alpha}{scale factor a of symmetric banded matrix A; default 1}

\item{x}{input vector; list of R external GPU pointer and dimension}

\item{A}{input matrix; list of R external GPU pointer and dimension}
}
\value{
updated matrix A
}
\description{
This function performs rank 1 update, A = a x x T + A, 
where A is symmetric matrix, x is vector, a is scalar
by using CUDA cublas function cublasDsyr
}
\seealso{
\code{\link{gergpu}}
}

