% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3blas.R
\name{syrkgpu}
\alias{syrkgpu}
\title{syrkgpu}
\usage{
syrkgpu(fillmode = 1, trans = 1, alpha = 1, A, beta = 0, C)
}
\arguments{
\item{fillmode}{indicates if matrix A lower or upper part is stored, 
the other part is not referenced and is inferred from the 
stored elements. if fillmode == 1 then the triagular 
banded matrix A is stored column by column, with the main diagonal of 
the matrix stored in row 1, the first subdiagonal in row 2 
(starting at first position), 
the second subdiagonal in row 3 (starting at first position), etc. 
if fillmode == 2 then the triangular banded matrix A is 
stored column by column, with the main diagonal of the matrix stored 
in row k+1, the first superdiagonal in row k (starting at second position),
the second superdiagonal in row k-1 (starting at third position), etc.}

\item{trans}{matrix A transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{C}{input/output matrix; list of R external GPU pointer and dimension}
}
\value{
updated matrix C, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix C's number of rows}
\item{n: }{matrix C's number of columns}
}
}
\description{
This function performs the symmetric rank- k update 
C = a op ( A ) op ( A ) T + b C
by using CUDA cublas function cublasDsyrk
}
\examples{
creategpu(1:12,3,4)->A
creategpu(1:12,4,3)->B
creategpu(1:9,3,3)->C
syrkgpu(fillmode=1,trans=1,alpha=1,A,beta = 0, C)
gathergpu(C)
}
\seealso{
\code{\link{gemmgpu}}
}

