% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{tbsvgpu}
\alias{tbsvgpu}
\title{tbsvgpu}
\usage{
tbsvgpu(fillmode = 1, trans = 1, diagmode = 1, k, A, x)
}
\arguments{
\item{fillmode}{indicates if matrix A lower or upper part is stored, 
the other symmetric part is not referenced and is inferred from the 
stored elements. if fillmode == 1 then the triagular 
banded matrix A is stored column by column, with the main diagonal of 
the matrix stored in row 1, the first subdiagonal in row 2 
(starting at first position), 
the second subdiagonal in row 3 (starting at first position), etc. 
if fillmode == 2 then the triangular banded matrix A is 
stored column by column, with the main diagonal of the matrix stored 
in row k+1, the first superdiagonal in row k (starting at second position),
the second superdiagonal in row k-1 (starting at third position), etc.}

\item{trans}{matrix A transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}

\item{diagmode}{indicates whether the main diagonal of the matrix A 
is unity and consequently should not be touched or modified by the function.
if diagmode = 1, the matrix diagonal has non-unit elements,
if diagmode = 2, the matrix diagonal has unit elements}

\item{k}{number of sub- or super- diagonals}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{x}{input/output vector; list of R external GPU pointer and dimension}
}
\value{
updated vector x, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{length of vector x}
\item{n: }{1}
}
}
\description{
This function solves the triangular banded linear system op(A) x = b
by using CUDA cublas function cublasDtbsv
}
\seealso{
\code{\link{tbmvgpu}}
}

