% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lapack.R
\name{tGPU}
\alias{tGPU}
\title{tGPU}
\usage{
tGPU(X)
}
\arguments{
\item{X}{input matrix; list of R external GPU pointer and dimension}
}
\value{
matrix transpose, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix X's number of rows}
\item{n: }{matrix X's number of columns}
}
}
\description{
This function transposes the given matrix 
by using CUDA cublas cublasDgeam
}
\examples{
a <- 1:12
a_gpu <- createGPUmat(a,3,4)
tGPU(a_gpu)->c_gpu
gatherGPU(c_gpu)
}
\seealso{
\code{\link{createGPUmat}}
}

