% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3blas.R
\name{trsmgpu}
\alias{trsmgpu}
\title{trsmgpu}
\usage{
trsmgpu(sidemode = 1, fillmode = 1, trans = 1, diagmode = 1,
  alpha = 1, A, B)
}
\arguments{
\item{sidemode}{indicates whether the given matrix is on the left or right side
in the matrix equation solved by a particular function. If sidemode == 1, 
the matrix is on the left side in the equation If sidemode == 2, 
the matrix is on the right side in the equation.}

\item{fillmode}{indicates if matrix A lower or upper part is stored, 
the other part is not referenced and is inferred from the 
stored elements. if fillmode == 1 then the triagular 
banded matrix A is stored column by column, with the main diagonal of 
the matrix stored in row 1, the first subdiagonal in row 2 
(starting at first position), 
the second subdiagonal in row 3 (starting at first position), etc. 
if fillmode == 2 then the triangular banded matrix A is 
stored column by column, with the main diagonal of the matrix stored 
in row k+1, the first superdiagonal in row k (starting at second position),
the second superdiagonal in row k-1 (starting at third position), etc.}

\item{trans}{matrix A transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}

\item{diagmode}{indicates whether the main diagonal of the matrix A 
is unity and consequently should not be touched or modified by the function.
if diagmode = 1, the matrix diagonal has non-unit elements,
if diagmode = 2, the matrix diagonal has unit elements.}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{B}{input/output matrix; list of R external GPU pointer and dimension}
}
\value{
updated matrix B, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix B's number of rows}
\item{n: }{matrix B's number of columns}
}
}
\description{
This function solves the triangle linear system 
A X = a B or X A = a B 
by using CUDA cublas function cublasDtrsm
}
\examples{
A_gpu <- creategpu(1:9, 3, 3)
B_gpu <- creategpu(1:6, 3, 2)
C_gpu <- creategpu(1:4, 3, 2)
symmgpu(alpha=1, A_gpu, B_gpu, beta=1, C_gpu)
gathergpu(C_gpu)
}
\seealso{
\code{\link{trmmgpu}}
}

