% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number_genertor.R
\name{uniformRNGGPU}
\alias{uniformRNGGPU}
\title{standard uniform random number generator}
\usage{
uniformRNGGPU(n, seed = 1)
}
\arguments{
\item{n}{number of random numbers}

\item{seed}{random number generator seed}
}
\value{
random numbers vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix X's number of rows}
\item{n: }{matrix X's number of columns}
}
}
\description{
This function generates uniformly distributed numbers between 0 and 1 
by using CUDA curand function CURAND_RNG_PSEUDO_DEFAULT and curandGenerateUniformDouble
}
\examples{
a_gpu <- uniformRNGGPU(100, 15)
gatherGPU(a_gpu)
}
\seealso{
\code{\link{createGPU}} \code{\link{createGPUmat}}
}

