% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_define_CUDA_functions.R
\name{vargpu}
\alias{vargpu}
\title{vargpu}
\usage{
vargpu(x)
}
\arguments{
\item{x}{list consisting of R external GPU pointer and dimension}
}
\value{
vector/matrix variance
}
\description{
Compute the variance of given vector/matrix
}
\details{
This function computes the variance of given vector/matrix 
by using self-defined CUDA function
}
\examples{
a <- creategpu(1:4)
vargpu(a)
}
\author{
Yuan Li
}
\seealso{
\code{\link{sumgpu}}
}
\keyword{CUDA}
\keyword{reduction}

