% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{addgpu}
\alias{addgpu}
\title{addgpu}
\usage{
addgpu(x, y)
}
\arguments{
\item{x}{list consisting of R external GPU pointer and dimension}

\item{y}{list consisting of R external GPU pointer and dimension}
}
\value{
element-wise addition of two vectors/matrices (x + y), 
a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the element-wise addition of two given 
vectors/matrices by using CUDA cublas function cublasDgeam
}
\examples{
a <- 1:4
b <- 2:5
a_gpu <- creategpu(a)
b_gpu <- creategpu(b)
addgpu(a_gpu, b_gpu) -> c_gpu
gathergpu(c_gpu)
}
\seealso{
\code{\link{subtractgpu}}
}

