% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGPU.R
\name{createGPU}
\alias{createGPU}
\title{createGPU}
\usage{
createGPU(input)
}
\arguments{
\item{input}{R vector to be copied (numerical)}
}
\value{
a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{vector length}
\item{n: }{always 1 as vector}
}
}
\description{
Create a GPU vector by copying from the input R vector
}
\details{
This function creates a vector in GPU by calling the CUDA cudamalloc function, and then copys
the values of vector from input R vector. The output of this 
function is a list consisting of the GPU pointer and vector length .
}
\note{
output is a R external GPU pointer and can only be used in Rcublas functions
}
\examples{
a <- 1:4
a_gpu <- createGPU(a)
gatherGPU(a_gpu) 
}
\author{
Yuan Li
}
\seealso{
\code{\link{gatherGPU}}
}
\keyword{GPU}
\keyword{create}
\keyword{vector}

