\name{cudaSetDoubleForDevice}
\alias{cudaSetDoubleForDevice}
\title{Converts a double argument to be executed on a device}
\description{ \code{d} - Double to convert}
\usage{cudaSetDoubleForDevice()}



\references{\url{http://docs.nvidia.com/cuda/cuda-driver-api/index.htm}}
\keyword{programming}
\concept{GPU}
