\name{cudaVersion}
\alias{cudaVersion}
\title{Determine the version of the CUDA SDK}
\description{
  This function allows us to query the version of the CUDA
  SDK - software development toolkit.
}
\usage{
cudaVersion()
}
\value{
  An integer vector of length 2, with elements named
  driver and runtime. 
}
\references{
}
\author{
Duncan Temple Lang
}
\seealso{}
\examples{
cudaVersion()
}
\keyword{programming}
