% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{dnormGPU}
\alias{dnormGPU}
\title{dnormGPU}
\usage{
dnormGPU(input, mean = 0, sd = 1)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}

\item{mean}{mean}

\item{sd}{standard deviation}
}
\value{
normal distribution density function vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computing the normal density function of input vector
}
\examples{
a <- 1:4
a_gpu <- createGPU(a)
dnormGPU(a_gpu)->b_gpu
gatherGPU(b_gpu)
}
\seealso{
\code{\link{expGPU}}
}

