% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1blas.R
\name{dotgpu}
\alias{dotgpu}
\title{dotgpu}
\usage{
dotgpu(x, y)
}
\arguments{
\item{x}{list consisting of R external GPU pointer and dimension}

\item{y}{list consisting of R external GPU pointer and dimension}
}
\value{
the resulting dot product
}
\description{
This function computes the dot product of two given vectors/matrix
by using CUDA cublas function cublasDdot
}
\examples{
a <- 1:4
b <- 2:5
a_gpu <- creategpu(a)
b_gpu <- creategpu(b)
dotgpu(a_gpu, b_gpu)
}
\seealso{
\code{\link{nrm2gpu}}
}

