% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_object_interface.R
\name{gathergpu}
\alias{gathergpu}
\title{gathergpu}
\usage{
gathergpu(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and its dimension}
}
\value{
R vector
}
\description{
Copy GPU matrix/vector to R vector
}
\details{
This function copys GPU vector/matrix to R vector

The output is always R vector, 
and GPU matrix will be copied by row-major. For example, an m by n
GPU matrix will be converted to a m*n R vector.
}
\note{
output is R vector and can be used by any R functions
}
\examples{
a <- 1:6
am_gpu <- createGPUmat(a, 3, 2)
gathergpu(am_gpu) 
}
\author{
Yuan Li
}
\seealso{
\code{\link{gathergpu}} \code{\link{creategpu}}
}
\keyword{GPU}

