% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{gbmvgpu}
\alias{gbmvgpu}
\title{gbmvgpu}
\usage{
gbmvgpu(trans = 1, kl, ku, alpha = 1, A, x, beta = 0, y)
}
\arguments{
\item{trans}{matrix A transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}

\item{kl}{number of subdiagonals}

\item{ku}{number of superdiagonals}

\item{alpha}{scale factor a of banded matrix A; default 1}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{x}{input vector; list of R external GPU pointer and dimension}

\item{beta}{scale factor b of vector y; default 0}

\item{y}{input/output vector; list of R external GPU pointer and dimension}
}
\value{
vector y, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{length of vector y}
\item{n: }{1}
}
}
\description{
This function computes banded matrix-vector multipication y = a A x + b y
by using CUDA cublas function cublasDgbmv
}
\seealso{
\code{\link{gergpu}}
}

