% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lapack.R
\name{geamgpu}
\alias{geamgpu}
\title{geamgpu}
\usage{
geamgpu(transa = 1, transb = 1, alpha = 1, A, B, beta = 0, C)
}
\arguments{
\item{transa}{matrix A transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}

\item{transb}{matrix B transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}

\item{alpha}{scale factor a of matrix A; default 1}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{B}{input matrix; list of R external GPU pointer and dimension}

\item{beta}{scale factor b of matrix B; default 0}

\item{C}{output matrix; list of R external GPU pointer and dimension}
}
\value{
updated matrix C, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix C's number of rows}
\item{n: }{matrix C's number of columns}
}
}
\description{
This function computes the matrix-matrix addition/trasportation 
C = a op ( A ) + b op ( B ) 
by using CUDA cublas function cublasDgeam
}
\seealso{
\code{\link{gemvgpu}}
}

