% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3blas.R
\name{gemmgpu}
\alias{gemmgpu}
\title{gemmgpu}
\usage{
gemmgpu(transa = 1, transb = 1, alpha = 1, A, B, beta = 0, C)
}
\arguments{
\item{transa}{matrix A transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}

\item{transb}{matrix B transpose operator, 1 (non-transpose), 2 (transpose),
3 (conjugate transpose); default at 1 (non-transpose)}

\item{alpha}{scale factor a of matrix A; default 1}

\item{A}{input matrix; list of R external GPU pointer and dimension}

\item{B}{input matrix; list of R external GPU pointer and dimension}

\item{beta}{scale factor b of matrix C; default 0}

\item{C}{input/output matrix; list of R external GPU pointer and dimension}
}
\value{
updated matrix C, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix C's number of rows}
\item{n: }{matrix C's number of columns}
}
}
\description{
This function computes the matrix-matrix multiplication 
C = a op ( A ) op ( B ) + b C 
by using CUDA cublas function cublasDgemm
}
\examples{
A_gpu <- creategpu(1:6, 3, 2)
B_gpu <- creategpu(1:6, 3, 2)
C_gpu <- creategpu(1:4, 2, 2)
gemmgpu(2, 1, 1, A_gpu, B_gpu, beta=1, C_gpu)
gathergpu(C_gpu)
}
\seealso{
\code{\link{gemvgpu}}
}

