% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu_object_class.R
\name{GPUobject}
\alias{GPUobject}
\title{GPUobject}
\usage{
GPUobject(input, length1, length2)
}
\arguments{
\item{input}{R external pointer}

\item{length1}{vector length}

\item{length2}{always 1 as vector}
}
\value{
a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{vector length}
\item{n: }{always 1 as vector}
}
}
\description{
classify the input as GPU vector and assign its dimension
}
\details{
This function classifies the input object as GPU vector and assign its dimension.
The output of this function is a list consisting of the GPU pointer and its dimension.
}
\note{
output is a R external GPU pointer and can only be used in Rcublas functions
}
\author{
Yuan Li
}
\seealso{
\code{\link{gatherGPU}}
}
\keyword{GPU}
\keyword{create}
\keyword{vector}

