% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviceQuery.R
\name{gpuQuery}
\alias{gpuQuery}
\title{gpuQuery}
\usage{
gpuQuery()
}
\description{
This function returns the information of available GPU device in system
}
\examples{
gpuQuery()
}
\seealso{
\code{\link{createGPU}}
}

