% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lapack.R
\name{inverseGPU}
\alias{inverseGPU}
\title{inverseGPU}
\usage{
inverseGPU(X)
}
\arguments{
\item{X}{input matrix; list of R external GPU pointer and dimension}
}
\value{
matrix inverse, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{matrix X's number of rows}
\item{n: }{matrix X's number of columns}
}
}
\description{
This function computes the inverse of given matrix (square) 
by using CUDA cublas function cublasDgetrfBatched and cublasDgetriBatched (LU decomposition)
}
\examples{
a <- 1:9
a_gpu <- createGPUmat(a,3,3)
inverseGPU(a_gpu)->c_gpu
gatherGPU(c_gpu)
}
\seealso{
\code{\link{mmGPU}} \code{\link{createGPUmat}}
}

