% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minGPU.R
\name{minGPU}
\alias{minGPU}
\title{minGPU}
\usage{
minGPU(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
the resulting index
}
\description{
finds the (smallest) index of the element with the minimum magnitude of given vector
This function finds the (smallest) index of the element with the minimum magnitude of given vector
by using CUDA cublas function cublasIdamin
}
\examples{
a <- 1:4
a_gpu <- createGPU(a)
minGPU(a_gpu) 
}
\seealso{
\code{\link{maxGPU}}
}

