% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1blas.R
\name{norm2GPU}
\alias{norm2GPU}
\title{norm2GPU}
\usage{
norm2GPU(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
vector Euclidean norm, a non-negative number.
}
\description{
Compute the Euclidean norm of given vector
}
\details{
This function computes Euclidean norm of given vector by using CUDA cublas function cublasDnrm2
}
\examples{
a <- 1:4
a_gpu <- createGPU(a)
norm2GPU(a_gpu) 
}
\author{
Yuan Li
}
\seealso{
\code{\link{gatherGPU}} \code{\link{createGPU}}
}
\keyword{GPU}

