% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnormGPU.R
\name{pnormGPU}
\alias{pnormGPU}
\title{pnormGPU}
\usage{
pnormGPU(input)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}
}
\value{
standard normal CDF result, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computing the standard normal cumulative density function input vector
}
\examples{
a <- 1:4
a_gpu <- createGPU(a)
pnormGPU(a_gpu)->b_gpu
gatherGPU(b_gpu)
}
\seealso{
\code{\link{qnormGPU}}
}

