% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_wise_math.R
\name{powergpu}
\alias{powergpu}
\title{powergpu}
\usage{
powergpu(input, alpha = 1)
}
\arguments{
\item{input}{list consisting of R external GPU pointer and dimension}

\item{alpha}{power factor}
}
\value{
powered vector/matrix, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the power of given vector/matrix 
by using self-defined CUDA function
}
\examples{
a <- 1:4
b <- 2
a_gpu <- creategpu(a)
powergpu(a_gpu, b) -> b_gpu
gathergpu(b_gpu)
}
\seealso{
\code{\link{sqrtgpu}}
}

