% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtGPU.R
\name{sqrtGPU}
\alias{sqrtGPU}
\title{sqrtGPU}
\usage{
sqrtGPU(input)
}
\arguments{
\item{input}{(non-negative vector) list consisting of R external GPU pointer and dimension}
}
\value{
square root vector, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the square root of given vector or matrix
by using CUDA function
}
\examples{
a <- 1:4
a_gpu <- createGPU(a)
sqrtGPU(a_gpu)->b_gpu
gatherGPU(b_gpu)
}
\seealso{
\code{\link{expGPU}}
}

