% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2blas.R
\name{subtractGPU}
\alias{subtractGPU}
\title{subtractGPU}
\usage{
subtractGPU(x, y)
}
\arguments{
\item{x}{list consisting of R external GPU pointer and dimension}

\item{y}{list consisting of R external GPU pointer and dimension}
}
\value{
subtraction of vectors or matrices (x - y), a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function computes the element-wise subtraction of two given vectors or matrices (x - y)
by using CUDA cublas function cublasDgeam
}
\examples{
a <- 1:4
b <- 2:5
a_gpu <- createGPU(a)
b_gpu <- createGPU(b)
subtractGPU(a_gpu, b_gpu)->c_gpu
gatherGPU(c_gpu)
}
\seealso{
\code{\link{addGPU}}
}

