% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lapack.R
\name{tgpu}
\alias{tgpu}
\title{tgpu}
\usage{
tgpu(X)
}
\arguments{
\item{X}{input matrix; list of R external GPU pointer and dimension}
}
\value{
matrix transpose, a list consisting of
\itemize{
\item{ptr: }{GPU pointer}
\item{m: }{number of rows}
\item{n: }{number of columns}
}
}
\description{
This function transposes the given matrix 
by using CUDA cublas cublasDgeam
}
\examples{
a <- 1:12
a_gpu <- creategpu(a, 3, 4)
tgpu(a_gpu) -> c_gpu
gathergpu(c_gpu)
}
\seealso{
\code{\link{creategpu}}
}

