% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varGPU.R
\name{varGPU}
\alias{varGPU}
\title{varGPU}
\usage{
varGPU(x)
}
\arguments{
\item{x}{list consisting of R external GPU pointer and dimension}
}
\value{
sample variance
}
\description{
Compute the variance of given sample
}
\details{
This function computes variance of given sample by using CUDA vector reduction
}
\examples{
a <- createGPU(1:4)
varGPU(a)
}
\author{
Yuan Li
}
\seealso{
\code{\link{sumGPU}}
}
\keyword{GPU}

